ia-services/requirements.txt
# Dépendances pour les services IA

# Embeddings et NLP
sentence-transformers>=2.2.2
transformers>=4.21.0
torch>=1.12.0
numpy>=1.21.0

# Vector store
faiss-cpu>=1.7.4
# Pour GPU: faiss-gpu>=1.7.4

# Parsing de documents
PyPDF2>=3.0.0
python-docx>=0.8.11

# Utilitaires
scikit-learn>=1.1.0
pandas>=1.4.0
